move hardcoding header/info out of Tutorial component
